Move configurations from config.py
Moved application configuration from config package to app.py
to allow for ease of finding configurations.

Resolves: Move configurations from config.py
