+++ .gitignore
@@ -108,4 +108,4 @@ venv.bak/
 
 .vscode
 Dockerfile
-*.log
+watch-folder
