--- app/logs/app_log.log
@@ -1,1 +1,0 @@
-
