--- config/config.py
@@ -1,45 +1,0 @@
-from os import path, getcwd
-import logging
-
-# FILE CONFIG
-BASE_PATH = getcwd()
-WATCH_PATH = path.join(BASE_PATH)
-WATCH_EXTS = [r".*\d{4,4}\-\d{2,2}\-\d{2,2}\-[A-Za-z]*\.pdf$"]
-
-# LOGGING CONFIG
-LOG_LOCATION = path.join(BASE_PATH, 'app','logs', 'app_log.log')
-DEFAULT_LOG_LEVEL = logging.INFO
-
-# SMTP SERVER CONFIG
-SMTP_SERVER = 'smtp.gmail.com'
-SMTP_PORT = 465  # SSL Port
-
-# EMAIL CONFIG
-EMAIL_ADDRESS = ''
-PASSWORD = ''
-MAIL_DATA_PATH = path.join(getcwd(),'app', 'assets', 'message_data', 'email', 'email.json')
-
-# Adding properties to config dictionary
-APP_CONFIG = dict()
-WATCHER_CONFIG = dict()
-LOGGING_CONFIG = dict()
-SMTP_CONFIG = dict()
-EMAIL_CONFIG = dict()
-
-APP_CONFIG['WATCHER_CONFIG'] = WATCHER_CONFIG
-APP_CONFIG['LOGGING_CONFIG'] = LOGGING_CONFIG
-APP_CONFIG['SMTP_CONFIG'] = SMTP_CONFIG
-APP_CONFIG['EMAIL_CONFIG'] = EMAIL_CONFIG
-
-WATCHER_CONFIG['WATCH_PATH'] = WATCH_PATH
-WATCHER_CONFIG['WATCH_EXTS'] = WATCH_EXTS
-
-LOGGING_CONFIG['LOG_LOCATION'] = LOG_LOCATION
-LOGGING_CONFIG['DEFAULT_LOG_LEVEL'] = DEFAULT_LOG_LEVEL
-
-SMTP_CONFIG['SMTP_SERVER'] = SMTP_SERVER
-SMTP_CONFIG['SMTP_PORT'] = SMTP_PORT
-
-EMAIL_CONFIG['EMAIL_ADDRESS'] = EMAIL_ADDRESS
-EMAIL_CONFIG['PASSWORD'] = PASSWORD
-EMAIL_CONFIG['MAIL_DATA_PATH'] = MAIL_DATA_PATH
